# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/notes/{id}":{"put":{"tags":["note-controller"],"operationId":"updateNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteDTO"}}}}}},"delete":{"tags":["note-controller"],"operationId":"deleteNote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/notes":{"get":{"tags":["note-controller"],"operationId":"getAllActiveNotes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NoteDTO"}}}}}}},"post":{"tags":["note-controller"],"operationId":"createNote","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteDTO"}}}}}}},"/api/notes/{id}/archive":{"patch":{"tags":["note-controller"],"operationId":"toggleArchive","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoteDTO"}}}}}}},"/api/notes/archived":{"get":{"tags":["note-controller"],"operationId":"getAllArchivedNotes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NoteDTO"}}}}}}}}},"components":{"schemas":{"NoteDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"archived":{"type":"boolean"}}}}}}